Consolidate plugin code to PluginRegistry class
